+++ src/config.ts
@@ -2,6 +2,7 @@ import { ZKTypeConfig } from "./types";
 
 export const defaultCircuitArtifactGeneratorConfig: ZKTypeConfig = {
   basePath: "circuits",
+  projectRoot: process.cwd(),
   circuitsASTPaths: [],
   outputArtifactsDir: "artifacts/circuits",
   outputTypesDir: "generated-types/circuits",