+++ test/helpers/generateTypes.ts
@@ -2,10 +2,13 @@ import { generateAST } from "./generator";
 
 import { CircuitTypesGenerator } from "../../src";
 
+import { findProjectRoot } from "../../src/utils";
+
 const astDir = "test/cache/circuits-ast";
 
 const circuitTypesGenerator = new CircuitTypesGenerator({
   basePath: "test/fixture",
+  projectRoot: findProjectRoot(process.cwd()),
   circuitsASTPaths: [
     "test/cache/circuits-ast/Basic.json",
     "test/cache/circuits-ast/credentialAtomicQueryMTPV2OnChainVoting.json",